openshift_checks aos_version: also check installed under yum
Tweaks to the logic around using yum vs dnf; now uses ansible_pkg_mgr to
determine which is in effect for a host.
Also, extended the yum logic to check installed packages in addition to
available packages in the aos_version module so that disconnected
installs and others with weird repo configs need not disable the
package_version check.