Restore registires to /etc/sysconfig/docker
Previously, a commit was added to migrate registires
from /etc/sysconfig/docker to /etc/containers/registries.conf

We are not currently enforcing a minimum version of docker
to consume from this new file, thus some installations
are not utilizing the correct repositories.

This commit duplicates the registires in both locations
to ensure additional/blocked/insecure registries are
honored.