Separate tuned daemon setup into a role.
Currently, profiles for the tuned daemon are set only for
OpenShift node(s).  This excludes the OpenShift loadbalancer.
As a result, ARP cache limits on loadbalancers are not raised.
This causes problems with HA setups where loadbalancers serve
1k+ OpenShift nodes.

This commit ensures the openshift-control-plane role is applied
to loadbalancers, masters and OpenShift infra nodes.  Regular
OpenShift worker nodes get the openshift-node profile.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1498213