Revert "Ensure minimum kernel version"
This reverts commit 2d54066fb149c6dcc38ed1be6529235ba1c3db92.

Conflicts:
	playbooks/init/base_packages.yml

We have clusters with nodes controlled by auto-scaling groups
whose kernel version does not meet this requirement.

We plan to address this by replacing said nodes with new nodes
instantiated from a newer image with an up-to-date kernel.

Unfortunately the "upgrade_scale_groups.yml" playbook includes the
"base_packages.yml" playbook after replacement nodes are scaled up,
and fails because the old nodes don't meet the version requirement.

To get past this conflict, we need a branch of openshift-ansible
with the kernel version requirement removed.